Refactor Context files into Provider files - and remove not using contexts
