--- src/providers/index.ts
@@ -1,2 +1,0 @@
-export * from './AppProvider';
-// export * from './ServiceModalProvider';
